refactor: Remove folder containing single file
